requirements.txt
# --- Core Frameworks (GPU Compatible Versions) ---
torch==2.1.2
torchaudio==2.1.2
numpy==1.22.0

# --- ML & TTS ---
transformers==4.36.2
TTS==0.22.0
scipy>=1.11.0
langid

# --- API & Server ---
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.6.0
pydantic-settings>=2.0.0

# --- gRPC & Contracts ---
grpcio>=1.62.0
grpcio-tools>=1.62.0
sentiric-contracts-py @ git+https://github.com/sentiric/sentiric-contracts.git@v1.11.3

# --- Monitoring & Logging ---
prometheus-fastapi-instrumentator>=6.1.0
python-json-logger>=2.0.7

# --- Utilities ---
python-dotenv
ffmpeg-python
defusedxml>=0.7.1